chore: Move fragments to fragment/
